+++ changelog.md
@@ -6,6 +6,7 @@ Core:
 
 - Allow to leave workspace
 - Allow to kick employee (Only for owner)
+- Browser notifications
 - Allow to create employee
 - Owner role for employee